fix: msw and jest config for LLM
