+++ apps/ledger-live-mobile/__tests__/server.ts
@@ -1,0 +1,6 @@
+import { http, HttpResponse } from "msw";
+import { setupServer } from "msw/node";
+import handlers from "./handlers";
+
+const server = setupServer(...handlers);
+export { server, http, HttpResponse };
